requirements.txt
zoomus
python-dotenv
requests
